backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
openai==1.3.5
pinecone-client==4.1.2
sentence-transformers
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
python-dotenv==1.0.0
httpx==0.25.2
anthropic
pydantic-settings
google-generativeai